requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-generativeai==0.3.1
python-dotenv==1.0.0
pymongo==4.6.1
motor==3.3.1
requests==2.31.0
pandas==2.1.3
pydantic==2.5.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
redis==5.0.1
aiohttp==3.9.1 